# FullstackTest/Itm8OfficeWeather.Client/OpenAPIs/OpenAPI.yaml
x-generator: NSwag v14.0.7.0 (NJsonSchema v11.0.0.0 (Newtonsoft.Json v13.0.0.0))
openapi: 3.0.0
info:
  title: Itm8 Office Weather API
  version: 1.0.0
servers:
  - url: http://localhost:5217
paths:
  /api/weather/offices:
    get:
      tags:
        - OfficeWeather
      operationId: OfficeWeather_GetWeatherData
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OfficeWeather"
        500:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
components:
  schemas:
    OfficeWeather:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        location:
          type: string
        coordinates:
          $ref: "#/components/schemas/GeoCoordinate"
        temperatureC:
          type: number
          format: double
        precipitation:
          $ref: "#/components/schemas/Precipitation"
        weatherSymbol:
          $ref: "#/components/schemas/WeatherSymbol"
        validTime:
          type: string
          format: date-time
    GeoCoordinate:
      type: object
      additionalProperties: false
      properties:
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
    Precipitation:
      type: object
      additionalProperties: false
      properties:
        category:
          $ref: "#/components/schemas/PrecipitationCategory"
        percent:
          type: integer
          format: int32
    PrecipitationCategory:
      type: integer
      description: ""
      x-enumNames:
        - NoPrecipitation
        - Snow
        - SnowAndRain
        - Rain
        - Drizzle
        - FreezingRain
        - FreezingDrizzle
      enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
    WeatherSymbol:
      type: integer
      description: ""
      x-enumNames:
        - ClearSky
        - NearlyClearSky
        - VariableCloudiness
        - HalfclearSky
        - CloudySky
        - Overcast
        - Fog
        - LightRainShowers
        - ModerateRainShowers
        - HeavyRainShowers
        - Thunderstorm
        - LightSleetShowers
        - ModerateSleetShowers
        - HeavySleetShowers
        - LightSnowShowers
        - ModerateSnowShowers
        - HeavySnowShowers
        - LightRain
        - ModerateRain
        - HeavyRain
        - Thunder
        - LightSleet
        - ModerateSleet
        - HeavySleet
        - LightSnowfall
        - ModerateSnowfall
        - HeavySnowfall
      enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
        - 9
        - 10
        - 11
        - 12
        - 13
        - 14
        - 15
        - 16
        - 17
        - 18
        - 19
        - 20
        - 21
        - 22
        - 23
        - 24
        - 25
        - 26
        - 27
    ProblemDetails:
      type: object
      additionalProperties:
        nullable: true
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true